Refactor project into REST API instead of MVC; introduce lombok; alter database script

+++ src/main/java/com/bekrenov/clinic/repository/AppointmentRepository.java
@@ -6,8 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.time.LocalDate;
 import java.util.List;
 
-public interface AppointmentRepository extends JpaRepository<Appointment, Integer> {
+public interface AppointmentRepository extends JpaRepository<Appointment, Long> {
 
-    List<Appointment> findAppointmentsByDoctor_IdAndAppointmentDate(int id, LocalDate date);
+    List<Appointment> findAppointmentsByDoctor_IdAndAppointmentDate(Long id, LocalDate date);
 
 }
